refactor(chat): move _generateResponse (~1.4k lines) into its own part file
The single biggest chunk of the god file: the core response-generation
orchestrator `_generateResponse` (speaker selection, the per-speaker realism eval
trigger, system-prompt/context assembly, streaming, post-gen needs/realism
wiring) moved verbatim into chat_service_generation.dart as a `part of`
extension. Zero behaviour change — pure relocation.

Kept whole rather than carved up: splitting a 1.4k-line hot-path method is a
behavioural risk better done as its own separately-verified pass. The four static
system-prompt constants it references (defaultGroupSystemPrompt,
observerModeSystemPrompt, defaultKoboldSystemPrompt, defaultApiSystemPrompt) are
now qualified with `ChatService.` as extensions require.

chat_service.dart 7,898 → 6,489 lines (10,506 → 6,489 across the effort, −38%).

flutter analyze clean; full unit suite (1348) + parity gate green.
