+++ lib/services/chat/chat_service_generation.dart
@@ -1,0 +1,1439 @@
+// Copyright (C) 2026 Front Porch AI
+// SPDX-License-Identifier: AGPL-3.0-or-later
+//
+// This file is part of Front Porch AI.
+//
+// Front Porch AI is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Front Porch AI is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with Front Porch AI. If not, see <https://www.gnu.org/licenses/>.
+
+part of '../chat_service.dart';
+
+/// The core response generation orchestrator (`_generateResponse`): speaker
+/// selection, the single per-speaker realism eval trigger (group path), system
+/// prompt + context assembly, streaming, and post-generation needs/realism
+/// wiring. Extracted verbatim from `chat_service.dart` (zero behaviour change) to
+/// shrink the god file — it is a single ~1.4k-line method, kept whole here
+/// rather than carved up (which would be a behavioural risk) until a careful,
+/// separately-verified decomposition is warranted.
+extension ChatServiceGeneration on ChatService {
+  Future<void> _generateResponse(
+    GenerationMode mode, {
+    CharacterCard? guestSpeaker,
+  }) async {
+    final epoch = ++_generationEpoch;
+    _isGenerating = true;
+    _generationProgress = 0.0;
+    _tokensGenerated = 0;
+    _maxTokens = _sessionGenSettings.resolveMaxLength(_storageService);
+    _generationStartTime = DateTime.now();
+    _isBuffering = true;
+    _generationPhase = GenerationPhase.preparing;
+    _prefillStartTime = null;
+    _lastPerfData = null;
+    _sentenceBuffer = '';
+    notifyListeners();
+
+    // Track original model for call mode swap/restore (needs to be outside try/catch)
+    String? _originalModelName;
+
+    try {
+      final userName = _userPersonaService.persona.name;
+
+      // Determine the speaking character first (needed for system prompt priority)
+      CharacterCard speakingCharacter;
+      if (guestSpeaker != null) {
+        // Scene Guest (Lite NPC) turn — stays in 1:1 (_activeGroup remains null),
+        // the guest speaks in its own bubble. Carries NO Realism/Needs work
+        // (see the guestSpeaker == null guards in the post-gen block below).
+        speakingCharacter = guestSpeaker;
+      } else if (_activeGroup != null) {
+        speakingCharacter =
+            (mode == GenerationMode.continue_ &&
+                _messages.isNotEmpty &&
+                !_messages.last.isUser)
+            ? _groupCharacters.firstWhere(
+                (c) => c.name == _messages.last.sender,
+                orElse: () => _pickNextGroupCharacter(),
+              )
+            : _pickNextGroupCharacter();
+      } else {
+        speakingCharacter = _activeCharacter!;
+      }
+
+      // SINGLE realism eval path (group trigger): the picked group member gets
+      // their per-turn eval here, after selection, as it always has. The 1:1
+      // host runs the SAME `_evaluateRealismForUpcomingSpeaker` from sendMessage
+      // instead (fresh turns only) so regen — which calls _generateResponse
+      // directly — does NOT re-evaluate and drift the host's realism. Lite Scene
+      // Guests (guestSpeaker != null) carry no realism.
+      if (guestSpeaker == null &&
+          _activeGroup != null &&
+          _realismActiveThisMode) {
+        await _evaluateRealismForUpcomingSpeaker(speakingCharacter);
+      }
+
+      // ── System prompt selection (Path B clean hierarchy) ──
+      // 1. Group-level system prompt (if set) — base for the whole group.
+      // 2. Per-character group override (if set for the speaker in this group) — appended.
+      // 3. Character's normal card system prompt (fallback if no group override for them).
+      // 4. (Later) Per-character Author's Note is injected separately with its own strength.
+      String systemPrompt;
+
+      if (_activeGroup != null && _activeGroup!.systemPrompt.isNotEmpty) {
+        systemPrompt = _activeGroup!.systemPrompt;
+      } else if (_activeGroup != null) {
+        systemPrompt = _observerMode
+            ? ChatService.observerModeSystemPrompt
+            : ChatService.defaultGroupSystemPrompt;
+      } else if (speakingCharacter.systemPrompt.isNotEmpty) {
+        systemPrompt = speakingCharacter.systemPrompt;
+      } else if (_storageService.generationSettings.systemPrompt.isNotEmpty) {
+        systemPrompt = _storageService.generationSettings.systemPrompt;
+      } else {
+        final isApi = _llmProvider != null && !_llmProvider!.isLocal;
+        systemPrompt = isApi
+            ? ChatService.defaultApiSystemPrompt
+            : ChatService.defaultKoboldSystemPrompt;
+      }
+
+      // Path B: When in a group, always attempt to layer the per-character group override
+      // (and card fallback) on top. A group prompt no longer completely hides per-char instructions.
+      if (_activeGroup != null) {
+        final groupCharPrompt = getSystemPromptForGroupCharacter(
+          speakingCharacter,
+        ).trim();
+        if (groupCharPrompt.isNotEmpty) {
+          systemPrompt +=
+              '\n\n[Group-specific instructions for ${speakingCharacter.name}]\n$groupCharPrompt';
+        } else if (speakingCharacter.systemPrompt.isNotEmpty) {
+          // Fallback to the character's own card prompt only if no group-specific override
+          systemPrompt +=
+              '\n\n[Specific instructions for ${speakingCharacter.name}]\n${speakingCharacter.systemPrompt.trim()}';
+        }
+      }
+
+      // In call mode, inject voice-specific instructions for natural conversation
+      if (_callMode &&
+          _storageService.sttSettings.callSystemPrompt.isNotEmpty) {
+        systemPrompt +=
+            '\n\n[Voice Call Mode] ${_storageService.sttSettings.callSystemPrompt}';
+      }
+
+      // Build Lorebook content (group + per-character, respecting inherit + group worlds)
+      String loreContent = '';
+      final activeLoreStrings = <String>{}; // Set for deduplication
+
+      final inherit = _activeGroup?.inheritCharacterLorebooks ?? true;
+
+      // Group-level lorebook (highest priority)
+      if (_activeGroup != null && _activeGroup!.groupLorebook.isNotEmpty) {
+        try {
+          final json = jsonDecode(_activeGroup!.groupLorebook);
+          final gl = Lorebook.fromJson(json as Map<String, dynamic>);
+          final active = gl.entries.where(
+            (e) => e.enabled && (e.isTriggered || e.constant),
+          );
+          activeLoreStrings.addAll(active.map((e) => e.content));
+        } catch (_) {}
+      }
+
+      // Group-level attached worlds
+      if (_activeGroup != null) {
+        for (final wid in _activeGroup!.worldIds) {
+          final world = _worldRepository.worlds
+              .where((w) => w.name == wid)
+              .firstOrNull;
+          if (world == null) continue;
+          final active = world.lorebook.entries.where(
+            (e) => e.enabled && (e.isTriggered || e.constant),
+          );
+          activeLoreStrings.addAll(active.map((e) => e.content));
+        }
+      }
+
+      // Per-character (only if inheriting or no group)
+      if (inherit || _activeGroup == null) {
+        final loreCharacters = _activeGroup != null
+            ? _groupCharacters
+            : [_activeCharacter!];
+        for (final ch in loreCharacters) {
+          if (ch.lorebook != null) {
+            final activeEntries = ch.lorebook!.entries.where(
+              (e) => e.enabled && (e.isTriggered || e.constant),
+            );
+            activeLoreStrings.addAll(activeEntries.map((e) => e.content));
+          }
+          for (final worldName in ch.worldNames) {
+            final world = _worldRepository.worlds
+                .where((w) => w.name == worldName)
+                .firstOrNull;
+            if (world == null) continue;
+            final activeWorldEntries = world.lorebook.entries.where(
+              (e) => e.enabled && (e.isTriggered || e.constant),
+            );
+            activeLoreStrings.addAll(activeWorldEntries.map((e) => e.content));
+          }
+        }
+      }
+
+      if (activeLoreStrings.isNotEmpty) {
+        loreContent = "Context Info:\n${activeLoreStrings.join('\n')}\n";
+      }
+
+      // Build persona block(s)
+      String personaBlock;
+      if (_activeGroup != null) {
+        personaBlock = _groupCharacters
+            .map((ch) {
+              final persona = _macroResolver.resolve(
+                _getEffectivePersonality(ch),
+                MacroContext(userName: userName, characterName: ch.name),
+                section: 'persona',
+              );
+              return "${ch.name}'s Persona: $persona";
+            })
+            .join('\n');
+      } else {
+        personaBlock =
+            "${speakingCharacter.name}'s Persona: ${_macroResolver.resolve(
+              _getEffectivePersonality(speakingCharacter),
+              MacroContext(userName: userName, characterName: speakingCharacter.name),
+              section: 'persona',
+            )}";
+      }
+
+      // User persona — inject user's self-description + learned facts
+      final userPersonaBlock = await _buildUserPersonaBlock(userName);
+
+      // Scenario — use group scenario override if set, else first character
+      final String rawScenario;
+      if (_activeGroup != null && _activeGroup!.scenario.isNotEmpty) {
+        rawScenario = _activeGroup!.scenario;
+      } else {
+        final scenarioChar = _activeGroup != null
+            ? _groupCharacters.first
+            : speakingCharacter;
+        rawScenario = _getEffectiveScenario(scenarioChar);
+      }
+      String scenario = rawScenario;
+      // A Scene Guest drops into the HOST's ongoing scene — it has no scenario
+      // of its own. Blank it here (prompt-only; the shared library card is never
+      // mutated, so a /join'd full character keeps its real scenario for when it
+      // is the host). This also self-heals legacy guests minted with the host's
+      // scenario baked in (the "model thinks the guest IS the host" bug).
+      if (guestSpeaker != null) scenario = '';
+
+      String suffix = "";
+
+      if (mode == GenerationMode.normal) {
+        suffix = "\n${speakingCharacter.name}:";
+      } else if (mode == GenerationMode.impersonate) {
+        suffix = "\n$userName:";
+      } else if (mode == GenerationMode.continue_) {
+        // Suffix will be set after history is built — see below
+        suffix = "";
+      }
+
+      // Build example dialogues block
+      String mesExampleBlock = '';
+      if (_activeGroup != null) {
+        final examples = _groupCharacters
+            .where((ch) => ch.mesExample.isNotEmpty)
+            .map(
+              (ch) => _macroResolver.resolve(
+                ch.mesExample,
+                MacroContext(userName: userName, characterName: ch.name),
+                section: 'mesExample',
+              ),
+            )
+            .toList();
+        if (examples.isNotEmpty) {
+          mesExampleBlock = '${examples.join('\n')}\n';
+        }
+      } else if (speakingCharacter.mesExample.isNotEmpty) {
+        mesExampleBlock = '${speakingCharacter.mesExample}\n';
+      }
+
+      // Build post-history instructions block
+      String postHistoryBlock = '';
+      if (_activeGroup == null &&
+          speakingCharacter.postHistoryInstructions.isNotEmpty) {
+        postHistoryBlock = '${speakingCharacter.postHistoryInstructions}\n';
+      }
+
+      // Author's note — placed right before the character speaks for maximum influence
+      String authorNoteBlock = '';
+      if (_authorNote.isNotEmpty) {
+        authorNoteBlock = _buildAuthorNoteBlock();
+      }
+
+      // Per-character Author's Note (group mode only): if the current speaker has
+      // a personal note, inject it using the same strength-modulated style.
+      // Falls back gracefully (no-op) if absent. Appended after any group-level note.
+      if (_activeGroup != null) {
+        final charNote = getAuthorNoteForGroupCharacter(speakingCharacter);
+        if (charNote.isNotEmpty) {
+          // Use per-character strength if set, otherwise fall back to group default
+          final s = getAuthorNoteStrengthForGroupCharacter(speakingCharacter);
+          final name = speakingCharacter.name;
+          String perCharBlock;
+          if (s <= 3) {
+            perCharBlock =
+                "[Author's Note (gentle suggestion for $name): $charNote]\n";
+          } else if (s <= 7) {
+            perCharBlock = "[Author's Note (for $name): $charNote]\n";
+          } else {
+            perCharBlock =
+                "[Author's Note (IMPORTANT for $name — apply immediately): $charNote]\n";
+          }
+          authorNoteBlock += perCharBlock;
+        }
+      }
+
+      // One-shot entrance directive (forked-in character) — hidden, consumed
+      // here so it influences only this generation and never persists.
+      if (_entranceDirective != null) {
+        authorNoteBlock += '[${_entranceDirective!}]\n';
+        _entranceDirective = null;
+      }
+
+      // ── Scene Guests (Lite NPCs) prompt injection (1:1 only) ───────────
+      // Guests speak for themselves in their own bubbles, so the primary must
+      // not fully voice/narrate them. A guest turn instead gets a short line
+      // grounding it as a visitor in the host's scene.
+      if (_activeGroup == null) {
+        final hostName = _activeCharacter?.name ?? 'the main character';
+        if (guestSpeaker != null) {
+          // A guest turn reuses the host's full transcript, so the identity
+          // switch must be unmistakable or the model conflates the guest with
+          // the host (confirmed even on strong API models). State plainly that
+          // everything above belongs to the host/user and the reply is ONLY the
+          // guest, and forbid voicing anyone else.
+          authorNoteBlock +=
+              '[SCENE GUEST TURN. You are now ${guestSpeaker.name}, who is '
+              'present in this scene — you are NOT $hostName and NOT $userName. '
+              'Everything written above was said and done by $hostName and '
+              '$userName; ${guestSpeaker.name} is a separate person. Reply ONLY '
+              'as ${guestSpeaker.name}: their own dialogue, actions, and '
+              'thoughts, reacting to what just happened. Do NOT write, speak, or '
+              'narrate anything for $hostName or $userName.]\n';
+        } else if (_sceneGuestCards.isNotEmpty) {
+          // Host turn with guests present: hard ban on ventriloquising them, or
+          // the host writes the guests' lines too (the "generated both at once"
+          // bug). Acknowledging/reacting is allowed; speaking for them is not.
+          final names = _sceneGuestCards.map((g) => g.name).join(', ');
+          authorNoteBlock +=
+              '[Also present in the scene: $names. Each of them speaks ONLY on '
+              'their own turn. Do NOT write any dialogue, actions, or inner '
+              'thoughts for them — not a single line. Stay entirely as '
+              '$hostName; you may have $hostName notice or react to them, but '
+              'never put words or actions on them.]\n';
+        }
+        // One-shot guest departure (armed by /exit) — narrated by the primary
+        // on this turn only, then cleared so it never persists.
+        if (guestSpeaker == null && _pendingGuestDeparture != null) {
+          authorNoteBlock +=
+              '[${_pendingGuestDeparture!} leaves the scene; '
+              'write them exiting naturally.]\n';
+          _pendingGuestDeparture = null;
+        }
+      }
+
+      // Build summary block if available
+      String summaryBlock = '';
+      if (_summary.isNotEmpty) {
+        summaryBlock = '[Summary of events so far: $_summary]\n';
+      }
+
+      // ── Continue mode: remove the last message from history ──
+      // For continue mode, we exclude the last message from the chat history
+      // and place it as the prompt suffix so the LLM continues from it naturally.
+      // Wrapped in try-finally to guarantee restoration even on exception.
+      ChatMessage? _continuePoppedMessage;
+      if (mode == GenerationMode.continue_ && _messages.isNotEmpty) {
+        _continuePoppedMessage = _messages.removeLast();
+        final partial = _continuePoppedMessage.text;
+        // For Continue: feed straight existing messages as the prompt (per user request).
+        // The suffix is the raw text of the message being continued (no re-added "Sender: " label).
+        // This makes the continuation prompt contain the plain previous messages + the exact
+        // partial text to extend, so the model continues the string directly without beginning
+        // the output with "Rachel:" or the speaker name.
+        // CRITICAL RULE: Strictly forbid the model from writing *anything* for {{user}} (actions, dialogue, thoughts, "he said", "you feel", etc.).
+        // This is a cardinal sin in AI RP. Only extend the provided partial text from the current speaker's POV and voice.
+        suffix =
+            "\n[CRITICAL RULE: The text below is an incomplete response from the *current speaker only*. You MUST ONLY generate more text that continues *this exact response* in the speaker's voice, style, and perspective. NEVER write any dialogue, actions, thoughts, narration, or descriptions for {{user}} or from {{user}}'s point of view. NEVER add new speaker labels or switch characters. Only append to the text below. Stop if it would require {{user}} content.]\n" +
+            partial;
+      }
+
+      // ── Macro resolution pass ──
+      final macroCtx = MacroContext(
+        userName: userName,
+        characterName: speakingCharacter.name,
+        summaryMaxWords: _storageService.memorySettings.summaryMaxWords,
+        chatId: _currentSessionId,
+        characterId: speakingCharacter.dbId,
+      );
+      systemPrompt = _macroResolver.resolve(
+        systemPrompt,
+        macroCtx,
+        section: 'systemPrompt',
+      );
+      if (loreContent.isNotEmpty) {
+        loreContent = _macroResolver.resolve(
+          loreContent,
+          macroCtx,
+          section: 'lore',
+        );
+      }
+      // personaBlock and group-mode examples are resolved per-character above
+      scenario = _macroResolver.resolve(
+        scenario,
+        macroCtx,
+        section: 'scenario',
+      );
+      if (_activeGroup == null && mesExampleBlock.isNotEmpty) {
+        mesExampleBlock = _macroResolver.resolve(
+          mesExampleBlock,
+          macroCtx,
+          section: 'mesExample',
+        );
+      }
+      if (postHistoryBlock.isNotEmpty) {
+        postHistoryBlock = _macroResolver.resolve(
+          postHistoryBlock,
+          macroCtx,
+          section: 'postHistory',
+        );
+      }
+
+      // Declare variables before try block so they're accessible after finally
+      String history = '';
+      String realismBlock = '';
+      String chanceTimeBlock = '';
+      String objectiveBlock = '';
+      String needsCatastropheBlock = '';
+      int droppedMessages = 0;
+
+      // Ensure the popped message is always restored, even if prompt assembly throws
+      try {
+        history = _buildChatHistory();
+
+        // ── Context Shift: budget-aware history trimming ──
+
+        // Realism / internal state block — now produced by a single dedicated composer
+        // (lib/services/chat/prompt_injection/realism_state_injection.dart).
+        // It groups *all* the live scalars (needs with x/100, bond/trust, emotion, time,
+        // arousal, spatial, etc.) under one clear, number-first header + collation guidance.
+        // This is the main place the model "sees" the current character state for consistency.
+        if (_realismActiveThisMode) {
+          realismBlock = _getRealismStateInjection();
+        }
+
+        // Chance Time injection — independent of realism mode
+        chanceTimeBlock = _getChanceTimeInjection();
+
+        // Objective injection — always injected regardless of realism mode
+        // Must sit in a fixed prompt section so it is NEVER trimmed by the budget system.
+        // (thin delegation to author_note_builder per step 8; state/CRUD in god)
+        objectiveBlock = _getObjectiveInjection();
+
+        // Mandatory Needs Catastrophe (Phase 2 stepping) — when a need hit 0 during
+        // the previous decay tick, we force the AI to roleplay the disaster right now.
+        if (_needsSimulation.pendingCatastrophe != null) {
+          needsCatastropheBlock =
+              '[MANDATORY CATASTROPHIC NEED EVENT — THIS HAS ALREADY OCCURRED THIS TURN:\n'
+              '${_needsSimulation.pendingCatastrophe}\n'
+              'You MUST narrate the immediate physical sensations, the visible evidence '
+              '(wet patch/puddle on clothes or floor, her collapsing or fainting, smell, '
+              'mortified/embarrassed expression, how {{user}} and anyone else present reacts), '
+              'and the emotional/social aftermath in the very first 1-2 paragraphs. '
+              'This is not optional, not a suggestion, and not something the character "might" do — '
+              'the event is canon and has just happened or is happening right now. '
+              'Do not fade to black, do not ask for permission, do not skip it.]\n';
+          // Consume it for this generation
+          _needsSimulation.consumePendingCatastrophe();
+        }
+
+        // Calculate token cost of all fixed sections to determine chat history budget
+        final fixedContent =
+            "$systemPrompt\n"
+            "$loreContent"
+            "$personaBlock\n"
+            "$userPersonaBlock"
+            "Scenario: $scenario\n"
+            "$mesExampleBlock"
+            "<START>\n"
+            "$summaryBlock"
+            "$postHistoryBlock"
+            "$authorNoteBlock"
+            "$objectiveBlock"
+            "$realismBlock"
+            "$needsCatastropheBlock"
+            "$suffix"
+            "$chanceTimeBlock";
+        final fixedTokens = await _countTokens(fixedContent);
+        final contextBudget = _sessionGenSettings.resolveContextSize(
+          _storageService,
+        );
+        final generationReserve =
+            _sessionGenSettings.resolveMaxLength(_storageService) +
+            50; // +50 safety margin
+        final historyBudget = contextBudget - fixedTokens - generationReserve;
+
+        if (historyBudget > 0) {
+          final result = await _buildChatHistoryWithBudget(historyBudget);
+          history = result.history;
+          droppedMessages = result.droppedCount;
+        }
+        // If budget is zero or negative, fixed sections already fill the context — use minimal history
+        if (historyBudget <= 0 && _messages.isNotEmpty) {
+          // Include at least the last message for continuity
+          final lastMsg = _messages.last;
+          history = lastMsg.characterId == '__director__'
+              ? '[Director: ${lastMsg.text}]'
+              : '${lastMsg.sender}: ${lastMsg.text}';
+          droppedMessages = _messages.length - 1;
+        }
+      } finally {
+        // ── Restore the popped continue message back into the list ──
+        if (_continuePoppedMessage != null) {
+          _messages.add(_continuePoppedMessage);
+        }
+      }
+
+      if (mode == GenerationMode.continue_) {
+        // Drop the needs/realism/relationship/chaos/objective/catastrophe state injections
+        // for Continue. Per user request: the continue prompt should be straight existing
+        // messages (the plain history transcript + the partial text to continue from).
+        // The runtime state blocks make the continuation feel injected and discordant.
+        realismBlock = '';
+        chanceTimeBlock = '';
+        objectiveBlock = '';
+        needsCatastropheBlock = '';
+        // Also skip RAG "earlier memories" for pure straight continuation.
+        droppedMessages = 0;
+      }
+
+      // ── RAG Memory Retrieval ──
+      // When messages are dropped from context, search for relevant past memories
+      // Skip retrieval for brand new chats to prevent old memories from interfering
+      String memoriesBlock = '';
+
+      final effectiveRagEnabled = _activeGroup != null
+          ? groupRagEnabled
+          : _storageService.memorySettings.ragEnabled;
+
+      if (_isNewChat) {
+        debugPrint(
+          '[RAG:Chat] Skipping memory retrieval - new chat in progress',
+        );
+      } else if (droppedMessages > 0 &&
+          _memoryService != null &&
+          effectiveRagEnabled) {
+        debugPrint(
+          '[RAG:Chat] ── Prompt assembly: $droppedMessages messages dropped, triggering retrieval ──',
+        );
+        try {
+          // Use last 3 messages as the query
+          final queryMessages = _messages.reversed
+              .take(3)
+              .map((m) => '${m.sender}: ${m.displayText}')
+              .join('\n');
+
+          // Scene Guests Phase 4: a guest turn retrieves the GUEST's own
+          // episodic memories (keyed on the guest id), not the host's. The
+          // injection format/budget below is shared — only the source id swaps.
+          final sourceIds = await _getMemorySourceIds(guest: guestSpeaker);
+          debugPrint('[RAG:Chat] Memory source IDs: $sourceIds');
+
+          final memories = await _memoryService!.retrieve(
+            queryText: queryMessages,
+            sourceCharacterIds: sourceIds,
+            currentSessionId: _currentSessionId ?? '',
+            inContextStart:
+                droppedMessages, // only search messages that are out of context
+            limit: groupRetrievalCount == 0 ? 9999 : groupRetrievalCount,
+            characterPriorities: currentGroupRAGPriorities,
+          );
+
+          if (memories.isNotEmpty) {
+            // Cap memory injection to the group's (or global) memory budget % of context.
+            // The summary carries the weight of context compression; RAG only
+            // supplements with specific details the summary missed. Too much
+            // RAG (2500+ tokens) overwhelms the model and causes it to
+            // reference stale events as if they're current ("going back in time").
+            final contextSize = _storageService.backendSettings.contextSize;
+            final budgetFraction = _activeGroup != null
+                ? (groupMemoryBudgetPercent / 100.0)
+                : 0.10;
+            final memoryBudget = (contextSize * budgetFraction).round();
+            final includedMemories = <String>[];
+            int usedTokens = 0;
+            for (final m in memories) {
+              final memTokens = (m.content.length / 4).ceil();
+              if (usedTokens + memTokens > memoryBudget &&
+                  includedMemories.isNotEmpty) {
+                debugPrint(
+                  '[RAG:Chat] ⚠ Trimmed ${memories.length - includedMemories.length} memories to fit budget ($memoryBudget tokens)',
+                );
+                break;
+              }
+              usedTokens += memTokens;
+              includedMemories.add('- ${m.content}');
+            }
+            if (includedMemories.isNotEmpty) {
+              memoriesBlock =
+                  '[Earlier in this conversation (already happened, do not revisit):\n${includedMemories.join('\n')}]\n';
+              debugPrint(
+                '[RAG:Chat] ✅ Injecting ${includedMemories.length}/${memories.length} memories (~$usedTokens tokens, budget: $memoryBudget)',
+              );
+            }
+          } else {
+            debugPrint('[RAG:Chat] No relevant memories found for this turn');
+          }
+        } catch (e) {
+          debugPrint('[RAG:Chat] ✗ RAG retrieval failed: $e');
+        }
+      } else if (droppedMessages > 0 &&
+          _storageService.memorySettings.ragEnabled) {
+        debugPrint(
+          '[RAG:Chat] ⚠ $droppedMessages messages dropped but RAG not operational (service=${_memoryService != null}, operational=${_memoryService?.isOperational ?? false})',
+        );
+      }
+
+      // Realism injection was already computed above for budget
+
+      // For chat APIs (OpenRouter, LM Studio), separate the system prompt
+      // so it can be sent as a proper 'system' role message.
+      final isRemoteApi = _llmProvider != null && !_llmProvider!.isLocal;
+      final chatSystemPrompt = isRemoteApi
+          ? "$systemPrompt\n$loreContent$personaBlock\n$userPersonaBlock"
+                "Scenario: $scenario\n$mesExampleBlock"
+          : null;
+
+      final prompt = isRemoteApi
+          ? "<START>\n"
+                "$summaryBlock"
+                "$memoriesBlock"
+                "$history"
+                "$postHistoryBlock"
+                "$authorNoteBlock"
+                "$objectiveBlock"
+                "$realismBlock"
+                "$needsCatastropheBlock"
+                "$suffix"
+                "$chanceTimeBlock"
+          : "$systemPrompt\n"
+                "$loreContent"
+                "$personaBlock\n"
+                "$userPersonaBlock"
+                "Scenario: $scenario\n"
+                "$mesExampleBlock"
+                "<START>\n"
+                "$summaryBlock"
+                "$memoriesBlock"
+                "$history"
+                "$postHistoryBlock"
+                "$authorNoteBlock"
+                "$objectiveBlock"
+                "$realismBlock"
+                "$needsCatastropheBlock"
+                "$suffix"
+                "$chanceTimeBlock";
+
+      // Track prompt budget for context viewer (always show full prompt)
+      _lastAssembledPrompt = chatSystemPrompt != null
+          ? '$chatSystemPrompt\n$prompt'
+          : prompt;
+      _lastPromptBudget = {
+        'System Prompt': (systemPrompt.length / 4).ceil(),
+        'Lorebook': (loreContent.length / 4).ceil(),
+        'Persona': (personaBlock.length / 4).ceil(),
+        'Scenario': ('Scenario: $scenario'.length / 4).ceil(),
+        'Examples': (mesExampleBlock.length / 4).ceil(),
+        'Summary': (summaryBlock.length / 4).ceil(),
+        'Retrieved Memories': (memoriesBlock.length / 4).ceil(),
+        'Chat History': (history.length / 4).ceil(),
+        'Post-History': (postHistoryBlock.length / 4).ceil(),
+        'Author\'s Note': (authorNoteBlock.length / 4).ceil(),
+        'Objectives': (objectiveBlock.length / 4).ceil(),
+        'Realism Mode': (realismBlock.length / 4).ceil(),
+        if (needsCatastropheBlock.isNotEmpty)
+          'Needs Catastrophe': (needsCatastropheBlock.length / 4).ceil(),
+        if (droppedMessages > 0) 'Dropped Messages': droppedMessages,
+      };
+      // Remove zero-value entries
+      _lastPromptBudget.removeWhere((_, v) => v == 0);
+
+      // Stop sequences: include character names, and user name (except when impersonating)
+      final g2 = _sessionGenSettings;
+      final stopSequences = {
+        ...g2.resolveStopSequences(_storageService).toSet(),
+      };
+
+      // In impersonate mode the model IS the user, so don't stop on user name
+      if (mode != GenerationMode.impersonate) {
+        stopSequences.add('\nUser:');
+        stopSequences.add('\n${_userPersonaService.persona.name}:');
+      }
+
+      // For Continue mode, do *not* stop on the current speaker's name.
+      // This lets the model produce long, natural extensions of the existing message
+      // in that character's voice without the name stop cutting it off mid-continuation.
+      // We still stop on other speakers or the user (to catch unwanted new turns).
+      String? continueSpeakerName;
+      if (mode == GenerationMode.continue_ &&
+          _messages.isNotEmpty &&
+          !_messages.last.isUser) {
+        continueSpeakerName = _messages.last.sender;
+      }
+
+      if (_activeGroup != null) {
+        for (final ch in _groupCharacters) {
+          if (continueSpeakerName != null && ch.name == continueSpeakerName) {
+            continue;
+          }
+          stopSequences.add('\n${ch.name}:');
+        }
+      } else {
+        final cur = _activeCharacter!.name;
+        if (continueSpeakerName == null || cur != continueSpeakerName) {
+          stopSequences.add('\n$cur:');
+        }
+      }
+      final stopList = stopSequences.toList();
+
+      // Get the active LLM service (local or remote)
+      final llmService =
+          testLlmServiceOverride ??
+          _llmProvider?.activeService ??
+          _koboldService;
+
+      // For call mode with a dedicated call model, temporarily swap the model
+      if (_callMode &&
+          _storageService.sttSettings.callModelName.isNotEmpty &&
+          _llmProvider != null &&
+          !_llmProvider!.isLocal) {
+        _originalModelName = _llmProvider!.openRouterService.modelName;
+        _llmProvider!.openRouterService.configure(
+          modelName: _storageService.sttSettings.callModelName,
+        );
+      }
+
+      final genParams = GenerationParams(
+        prompt: prompt,
+        systemPrompt: chatSystemPrompt,
+        maxLength: g2.resolveMaxLength(_storageService),
+        minLength: g2.resolveMinLength(_storageService),
+        minP: g2.resolveMinP(_storageService),
+        temperature: g2.resolveTemperature(_storageService),
+        repeatPenalty: g2.resolveRepeatPenalty(_storageService),
+        repPenTokens: g2.resolveRepeatPenaltyTokens(_storageService),
+        dynatempRange: g2.resolveDynamicTempEnabled(_storageService)
+            ? g2.resolveDynamicTempRange(_storageService)
+            : null,
+        xtcThreshold: g2.resolveXtcThreshold(_storageService),
+        xtcProbability: g2.resolveXtcProbability(_storageService),
+        stopSequences: stopList,
+        reasoningEnabled: (_callMode || mode == GenerationMode.continue_)
+            ? false
+            : g2.resolveReasoningEnabled(_storageService),
+        reasoningEffort: g2.resolveReasoningEffort(_storageService),
+        // Force zero thinking budget on Continue (and call mode) for providers like OpenRouter/Nano-GPT.
+        // This tells supported models (Kimi K2 Thinking, DeepSeek hybrid reasoning models, certain Qwen3 etc.)
+        // to spend 0 tokens on internal reasoning and answer directly, preventing the model from dumping
+        // its next analysis/think block into the visible character response.
+        reasoningMaxTokens: (_callMode || mode == GenerationMode.continue_)
+            ? 0
+            : null,
+        bannedPhrases: g2.resolveBannedPhrases(_storageService).isNotEmpty
+            ? g2.resolveBannedPhrases(_storageService)
+            : null,
+      );
+
+      // Get streaming response from whichever backend is active
+      final stream = llmService.generateStream(genParams);
+
+      // ── Phase: Prefilling ──
+      // The HTTP request is now in flight. For KoboldCPP, the model is
+      // processing the prompt (prefill/eval). No tokens arrive until
+      // prefill finishes. Poll /api/extra/perf for real-time status.
+      _generationPhase = GenerationPhase.prefilling;
+      _prefillStartTime = DateTime.now();
+      _prefillPromptTokens = (prompt.length / 4).ceil(); // Rough placeholder
+      notifyListeners();
+
+      // If using local KoboldCPP, poll /api/extra/perf during prefill
+      // to get real prompt processing metrics.
+      Timer? _perfPoller;
+      final isLocalBackend = _llmProvider == null || _llmProvider!.isLocal;
+      if (isLocalBackend) {
+        // Get REAL token count from the model's tokenizer (async, updates UI when done)
+        _koboldService.countTokens(prompt).then((realCount) {
+          if (_generationPhase == GenerationPhase.prefilling && realCount > 0) {
+            _prefillPromptTokens = realCount;
+            debugPrint(
+              '[Prefill] Actual token count from tokenizer: $realCount (was ~${(prompt.length / 4).ceil()} est)',
+            );
+            notifyListeners();
+          }
+        });
+
+        _perfPoller = Timer.periodic(const Duration(seconds: 2), (_) async {
+          if (_generationPhase != GenerationPhase.prefilling) {
+            _perfPoller?.cancel();
+            _perfPoller = null;
+            return;
+          }
+          final perf = await _koboldService.fetchPerf();
+          if (perf != null) {
+            _lastPerfData = perf;
+            notifyListeners();
+          }
+        });
+      }
+
+      String accumulatedResponse = "";
+      bool stopFound = false;
+      _tokenBuffer.clear();
+      _displayedTokenCount = 0;
+      _tokenTimestamps.clear();
+      bool streamDone = false;
+      DateTime? _thinkStartTime;
+      bool _thinkStarted = false;
+      bool _thinkEnded = false;
+
+      // Determine message identity
+      String originalText = '';
+      String targetSender;
+      bool isUserTarget;
+
+      if (mode == GenerationMode.continue_) {
+        originalText = _messages.last.text;
+        targetSender = _messages.last.sender;
+        isUserTarget = _messages.last.isUser;
+        // Merge metadata if continuing
+        if (_pendingRealismMetadata != null) {
+          _messages.last.activeMetadata ??= {};
+          _messages.last.activeMetadata!.addAll(_pendingRealismMetadata!);
+          _pendingRealismMetadata = null;
+        }
+      } else {
+        targetSender = mode == GenerationMode.normal
+            ? speakingCharacter.name
+            : _userPersonaService.persona.name;
+        isUserTarget = mode == GenerationMode.impersonate;
+        // A Scene Guest turn carries NO Realism/Needs, so its message must never
+        // inherit _pendingRealismMetadata — which still holds the HOST turn's
+        // verification result (the leftover "✓ Director accepted" chip), bond
+        // deltas, etc. Guests get clean (null) metadata.
+        final initialMetadata =
+            (guestSpeaker != null || _pendingRealismMetadata == null)
+            ? null
+            : Map<String, dynamic>.from(_pendingRealismMetadata!);
+        debugPrint(
+          '[Realism:Metadata] Attaching to new message: bond_delta=${initialMetadata?['bond_delta']}, keys=${initialMetadata?.keys.toList()}',
+        );
+        _messages.add(
+          ChatMessage(
+            text: "",
+            sender: targetSender,
+            isUser: isUserTarget,
+            characterId: mode == GenerationMode.normal
+                ? _getCharacterIdForCard(speakingCharacter)
+                : null,
+            metadata: initialMetadata,
+            swipeMetadata: initialMetadata != null ? [initialMetadata] : null,
+          ),
+        );
+        _pendingRealismMetadata = null;
+      }
+
+      // Helper to update the visible message from buffer
+      void _flushBufferToDisplay() {
+        if (epoch != _generationEpoch) return; // stale generation
+        if (_tokenBuffer.isEmpty && _displayedTokenCount == 0) return;
+        // Build displayed text from all tokens up to _displayedTokenCount
+        final displayTokens = _tokenBuffer.take(_displayedTokenCount).join();
+        String displayText;
+        if (mode == GenerationMode.continue_) {
+          displayText = originalText + displayTokens;
+        } else {
+          displayText = displayTokens.trimLeft();
+        }
+        // CRITICAL: Modify existing message in place to preserve thinkingStartTime and other metadata
+        _messages.last.text = displayText;
+        notifyListeners();
+      }
+
+      // Read display buffer settings — disable for remote APIs (they're fast enough)
+      final isRemoteBackend = _llmProvider != null && !_llmProvider!.isLocal;
+      final bufferEnabled = isRemoteBackend
+          ? false
+          : _storageService.uiSettings.displayBufferEnabled;
+      final targetTps = _storageService.uiSettings.targetDisplayTps;
+
+      // Drain timer: displays tokens at the user-configured constant rate
+      void _startDrainTimer() {
+        if (_drainTimer != null) return;
+        final interval = Duration(milliseconds: (1000.0 / targetTps).round());
+        _drainTimer = Timer.periodic(interval, (_) {
+          if (epoch != _generationEpoch) {
+            _drainTimer?.cancel();
+            _drainTimer = null;
+            return;
+          } // stale
+          if (_displayedTokenCount < _tokenBuffer.length) {
+            _displayedTokenCount++;
+            _flushBufferToDisplay();
+          } else if (streamDone) {
+            // Stream finished and buffer fully drained
+            _drainTimer?.cancel();
+            _drainTimer = null;
+          }
+          // If buffer is caught up but stream still running, timer ticks idly until more tokens arrive
+        });
+      }
+
+      // Consume the stream — tokens go into buffer (or display immediately)
+      await for (final token in stream) {
+        if (_cancelRequested) break;
+        accumulatedResponse += token;
+        _tokensGenerated++;
+        _tokenTimestamps.add(DateTime.now());
+
+        // ── Phase transition: first token marks end of prefill ──
+        if (_tokensGenerated == 1) {
+          _perfPoller?.cancel();
+          _perfPoller = null;
+          // Fetch final perf data so we know how long prefill really took
+          if (isLocalBackend) {
+            _koboldService.fetchPerf().then((perf) {
+              if (perf != null) {
+                _lastPerfData = perf;
+              }
+            });
+          }
+          _prefillStartTime = null;
+        }
+
+        // Broadcast token to external listeners (SSE bridge)
+        _tokenBroadcast.add(token);
+        _generationProgress = _maxTokens > 0
+            ? (_tokensGenerated / _maxTokens).clamp(0.0, 1.0)
+            : 0.0;
+
+        // Sentence streaming: accumulate tokens and emit complete sentences
+        _sentenceBuffer += token;
+
+        // Split strategy:
+        // 1. Always split at sentence boundaries: . ! ? followed by space, or \n
+        // 2. For long buffers (>80 chars / ~15 words), also split at clause
+        //    boundaries: ", " "; " " — " " - " to keep TTS chunks short (~1-3s)
+        bool emitted = true;
+        while (emitted) {
+          emitted = false;
+
+          // First try sentence boundaries
+          final sentenceEnd = RegExp(r'[.!?]\s|[.!?]$|\n');
+          if (sentenceEnd.hasMatch(_sentenceBuffer)) {
+            final match = sentenceEnd.firstMatch(_sentenceBuffer)!;
+            final sentence = _sentenceBuffer.substring(0, match.end).trim();
+            _sentenceBuffer = _sentenceBuffer.substring(match.end);
+            if (sentence.isNotEmpty) {
+              _sentenceBroadcast.add(sentence);
+              emitted = true;
+            }
+            continue;
+          }
+
+          // For long buffers, split at clause boundaries to keep TTS fast
+          if (_sentenceBuffer.length > 80) {
+            final clauseEnd = RegExp(r',\s|;\s|\s[—–-]\s');
+            if (clauseEnd.hasMatch(_sentenceBuffer)) {
+              // Find the LAST clause boundary to maximize chunk size
+              Match? lastMatch;
+              for (final m in clauseEnd.allMatches(_sentenceBuffer)) {
+                if (m.start > 30) lastMatch = m; // at least 30 chars per chunk
+              }
+              if (lastMatch != null) {
+                final chunk = _sentenceBuffer
+                    .substring(0, lastMatch.end)
+                    .trim();
+                _sentenceBuffer = _sentenceBuffer.substring(lastMatch.end);
+                if (chunk.isNotEmpty) {
+                  _sentenceBroadcast.add(chunk);
+                  emitted = true;
+                }
+              }
+            }
+          }
+        }
+
+        // Client-side safety trim check (mid-stream)
+        for (final stop in stopList) {
+          if (accumulatedResponse.contains(stop)) {
+            int index = accumulatedResponse.indexOf(stop);
+            final trimmedTotal = accumulatedResponse.substring(0, index);
+            final previousTotal = _tokenBuffer.join();
+            final lastTokenContribution = trimmedTotal.substring(
+              previousTotal.length.clamp(0, trimmedTotal.length),
+            );
+            if (lastTokenContribution.isNotEmpty) {
+              _tokenBuffer.add(lastTokenContribution);
+            }
+            accumulatedResponse = trimmedTotal;
+            stopFound = true;
+            break;
+          }
+        }
+
+        if (!stopFound) {
+          _tokenBuffer.add(token);
+        }
+
+        // Track think timing
+        if (!_thinkStarted && accumulatedResponse.contains('<think>')) {
+          _thinkStarted = true;
+          _thinkStartTime = DateTime.now();
+          _generationPhase = GenerationPhase.thinking;
+          if (_messages.isNotEmpty) {
+            _messages.last.thinkingStartTime =
+                _thinkStartTime.millisecondsSinceEpoch;
+          }
+        }
+        if (_thinkStarted &&
+            !_thinkEnded &&
+            accumulatedResponse.contains('</think>')) {
+          _thinkEnded = true;
+          // Transition out of thinking to buffering/generating
+          _generationPhase = bufferEnabled
+              ? GenerationPhase.buffering
+              : GenerationPhase.generating;
+          if (_thinkStartTime != null && _messages.isNotEmpty) {
+            _messages.last.thinkingDurationMs = DateTime.now()
+                .difference(_thinkStartTime)
+                .inMilliseconds;
+            // Keep thinkingStartTime for fallback display logic in UI
+          }
+        }
+        // If no thinking involved, first token transitions directly
+        if (!_thinkStarted && _tokensGenerated == 1) {
+          _generationPhase = bufferEnabled
+              ? GenerationPhase.buffering
+              : GenerationPhase.generating;
+        }
+
+        if (bufferEnabled) {
+          // Calculate current rolling TPS (last 3 seconds)
+          final now = DateTime.now();
+          final cutoff = now.subtract(const Duration(seconds: 3));
+          final recentCount = _tokenTimestamps
+              .where((t) => t.isAfter(cutoff))
+              .length;
+          final windowStart =
+              _tokenTimestamps.where((t) => t.isAfter(cutoff)).firstOrNull ??
+              _generationStartTime!;
+          final windowElapsed =
+              now.difference(windowStart).inMilliseconds / 1000.0;
+          final currentTps = (recentCount >= 2 && windowElapsed > 0)
+              ? recentCount / windowElapsed
+              : (_tokensGenerated > 0
+                    ? _tokensGenerated /
+                          (now
+                                  .difference(_generationStartTime!)
+                                  .inMilliseconds /
+                              1000.0)
+                    : 0.0);
+
+          if (_drainTimer == null && _tokensGenerated >= 10) {
+            // Not yet draining — calculate when to start
+            // Buffer target = how many tokens fill the configured duration
+            final bufferDuration =
+                _storageService.uiSettings.bufferDurationSeconds;
+            int bufferTarget;
+            if (currentTps > 0) {
+              bufferTarget = (currentTps * bufferDuration).round().clamp(
+                5,
+                _maxTokens,
+              );
+            } else {
+              bufferTarget = 30; // Fallback if TPS unknown
+            }
+
+            if (_tokenBuffer.length >= bufferTarget) {
+              _isBuffering = false;
+              _generationPhase = GenerationPhase.generating;
+              _startDrainTimer();
+            }
+          } else if (_drainTimer != null) {
+            // Already draining — check if buffer is running low
+            final remaining = _tokenBuffer.length - _displayedTokenCount;
+            if (remaining <= 3 && !streamDone) {
+              // Buffer critically low — pause drain to rebuild
+              _drainTimer?.cancel();
+              _drainTimer = null;
+              _isBuffering = true;
+              _generationPhase = GenerationPhase.buffering;
+            }
+          }
+        } else {
+          // No buffer: display tokens immediately
+          _isBuffering = false;
+          _generationPhase = GenerationPhase.generating;
+          _displayedTokenCount = _tokenBuffer.length;
+          _flushBufferToDisplay();
+        }
+
+        // Update TPS/progress in the bar even during buffering
+        notifyListeners();
+
+        if (stopFound) break;
+      }
+
+      // Mark stream as done
+      streamDone = true;
+      _isBuffering = false;
+
+      if (!bufferEnabled) {
+        // No buffer: everything already displayed
+        _displayedTokenCount = _tokenBuffer.length;
+        _flushBufferToDisplay();
+      } else if (_drainTimer == null) {
+        // Buffer never started draining (genTps < targetTps) — start now with all tokens ready
+        _startDrainTimer();
+        // Wait for drain to complete
+        while (_displayedTokenCount < _tokenBuffer.length) {
+          await Future.delayed(const Duration(milliseconds: 16));
+        }
+        _drainTimer?.cancel();
+        _drainTimer = null;
+      } else {
+        // Drain already running — wait for it to finish
+        while (_displayedTokenCount < _tokenBuffer.length) {
+          await Future.delayed(const Duration(milliseconds: 16));
+        }
+        _drainTimer?.cancel();
+        _drainTimer = null;
+      }
+
+      _isGenerating = false;
+      _cancelRequested = false;
+      _generationProgress = 0.0;
+      _isBuffering = false;
+      _generationPhase = GenerationPhase.idle;
+      _prefillStartTime = null;
+      _prefillPromptTokens = 0;
+      _generationStartTime = null;
+      _perfPoller?.cancel();
+      _perfPoller = null;
+
+      // Fetch final perf stats from KoboldCPP for post-generation display
+      if (isLocalBackend) {
+        _koboldService.fetchPerf().then((perf) {
+          if (perf != null) _lastPerfData = perf;
+        });
+      }
+
+      // Signal generation complete to SSE listeners
+      _tokenBroadcast.add('__DONE__');
+
+      // Flush remaining sentence buffer and signal done to sentence listeners
+      if (_sentenceBuffer.trim().isNotEmpty) {
+        _sentenceBroadcast.add(_sentenceBuffer.trim());
+        _sentenceBuffer = '';
+      }
+      _sentenceBroadcast.add('__DONE__');
+
+      notifyListeners();
+
+      // Only finalize if this generation is still current
+      if (epoch == _generationEpoch) {
+        String finalResponse = accumulatedResponse.trim();
+
+        // SillyTavern-like safety net for Continue (and call mode): even after requesting
+        // enabled:false + max_tokens:0 + exclude:true on the provider, some thinking models
+        // (Kimi 2.6:thinking etc.) can still emit stray <think> or reasoning text.
+        // Strip it from the final text before it becomes part of the character's message.
+        // This matches ST's "Strip Reasoning Tags" behavior as a client-side backstop.
+        if (mode == GenerationMode.continue_ || _callMode) {
+          finalResponse = _stripThinkBlocks(finalResponse);
+        }
+
+        // Snapshot which entries were already triggered before scanning the AI response.
+        // We will only decrement those — newly AI-triggered entries must keep their
+        // full depth budget so they are visible on the next user turn.
+        final preAiTriggered = <LorebookEntry>{};
+        final charactersForSnapshot = _activeGroup != null
+            ? _groupCharacters
+            : (_activeCharacter != null
+                  ? [_activeCharacter!]
+                  : <CharacterCard>[]);
+        for (final ch in charactersForSnapshot) {
+          if (ch.lorebook != null) {
+            for (final e in ch.lorebook!.entries) {
+              if (e.isTriggered && !e.constant) preAiTriggered.add(e);
+            }
+          }
+          for (final worldName in ch.worldNames) {
+            final world = _worldRepository.worlds
+                .where((w) => w.name == worldName)
+                .firstOrNull;
+            if (world == null) continue;
+            for (final e in world.lorebook.entries) {
+              if (e.isTriggered && !e.constant) preAiTriggered.add(e);
+            }
+          }
+        }
+
+        if (finalResponse.isNotEmpty) {
+          _lorebookScanner.scanLorebook(finalResponse);
+        }
+
+        // Decrement only entries that were active before the AI response.
+        // This preserves full depth for lore discovered in the AI's own words.
+        // Thin delegation (preAi set computed in god for snapshot; scanner owns decrement).
+        _lorebookScanner.decrementLoreDepthForEntries(preAiTriggered);
+
+        // Save session after AI message is complete
+        await _saveChat();
+
+        // ── Scene Guest (Lite NPC) parity guard ──────────────────────────
+        // A guest turn must NOT touch the active character's Realism Engine,
+        // Needs simulation, inter-character feelings, time, chips, or the
+        // periodic (facts/evolution/summary/RAG) evaluators. The guest carries
+        // no such state. Everything from here through the periodic evals is
+        // gated so guest presence/turns leave the primary's state untouched.
+        // (Lorebook scan + _saveChat above still ran for the guest.)
+        if (guestSpeaker == null) {
+          // Phase 2: Update hidden inter-character feelings for the speaker who
+          // just responded, based on what was said in the recent exchange.
+          // This makes the invisible tracking react to actual dialogue.
+          if (_activeGroup != null &&
+              !_observerMode &&
+              finalResponse.isNotEmpty) {
+            final lastSpeaker = _messages.isNotEmpty
+                ? _messages.last.sender
+                : '';
+            final speakerCard = _groupCharacters.firstWhere(
+              (c) => c.name == lastSpeaker,
+              orElse: () => _groupCharacters.first,
+            );
+            final speakerId = _getCharacterIdFromCard(speakerCard);
+            if (speakerId.isNotEmpty) {
+              _relationshipService
+                  .updateInterCharacterFeelingsFromRecentExchange(speakerId);
+              // (old checkpoint call removed in v30) // persist the hidden relationship changes
+            }
+          }
+
+          // For group non-observer turns, temporarily re-impersonate the speaker of the *just generated*
+          // response so the post-gen needs checks (now _runPostGenNeedsChecks thin to
+          // _needsImpactEvaluator) use the correct _activeCharacter (for name, personality/stance
+          // in the consolidated needs impact prompt). The pre-speaker-eval left the *scalars*
+          // (incl. needs vector) loaded for this speaker but restored the _activeCharacter pointer
+          // to the prior speaker; the thin delegate relies on the pointer for cbs. We restore the
+          // pointer after the checks (scalars remain correct for the persist below).
+          CharacterCard? prePostActiveChar;
+          if (_activeGroup != null && !_observerMode) {
+            prePostActiveChar = _activeCharacter;
+            _activeCharacter = speakingCharacter;
+            final sid = _getCharacterIdFromCard(speakingCharacter);
+            if (sid.isNotEmpty) {
+              _loadGroupRealismIntoScalars(sid);
+            }
+          }
+
+          await _runPostGenNeedsChecks(finalResponse);
+
+          if (prePostActiveChar != null) {
+            _activeCharacter = prePostActiveChar;
+          }
+
+          // For group non-observer, persist the post-scene + long-gen-decay needs changes (and any
+          // other scalars mutated by the checks) back into _groupRealism for this speaker. This is
+          // what makes sidebar member cards + getNeedsForGroupCharacter() + future loads see the
+          // effects of the just-generated response. (Pre-eval saved the pre-turn state for bond/etc;
+          // this captures the *response* effects on needs.)
+          if (_activeGroup != null &&
+              !_observerMode &&
+              finalResponse.isNotEmpty &&
+              _messages.isNotEmpty) {
+            final lastSender = _messages.last.sender;
+            final speakerCard = _groupCharacters.firstWhere(
+              (c) => c.name == lastSender,
+              orElse: () => _groupCharacters.first,
+            );
+            final sid = _getCharacterIdFromCard(speakerCard);
+            if (sid.isNotEmpty) {
+              _saveScalarsIntoGroupRealism(sid);
+            }
+          }
+
+          // Check if summary needs updating (fire-and-forget)
+          // Group name resolution for {{char}} in summary prompt is best-effort at trigger time (after prePostActiveChar restore dance); correct for 1:1, may use restored active or group fallback in group non-obs (timing-dependent per group impersonation; dispatch preserved via cbs). See leaf header + test for qualify.
+          _maybeUpdateSummary();
+
+          // Embed messages for RAG memory (fire-and-forget)
+          _maybeEmbedMessages();
+
+          // Periodic evaluations coordinator (facts + character evolution).
+          // Each now respects its own interval (autoPersonaInterval / evolutionInterval)
+          // via dedicated god-owned counters. Sequenced here when they coincide.
+          _maybeRunPeriodicEvals();
+        } // end Scene Guest parity guard (guestSpeaker == null)
+
+        // (Task completion check now runs pre-generation in sendMessage)
+
+        // TTS auto-play: speak the new character message automatically
+        if (_ttsService != null &&
+            _storageService.ttsSettings.ttsEnabled &&
+            _storageService.ttsSettings.ttsAutoPlay &&
+            _messages.isNotEmpty &&
+            !_messages.last.isUser) {
+          final lastMsg = _messages.last;
+          final msgId = 'msg_${_messages.length - 1}';
+          // Resolve per-character voice, falling back to global default
+          String? voiceKey;
+          if (_activeGroup != null) {
+            final charMatch = _groupCharacters
+                .where((c) => c.name == lastMsg.sender)
+                .firstOrNull;
+            voiceKey = charMatch?.ttsVoice;
+          } else {
+            voiceKey = _activeCharacter?.ttsVoice;
+          }
+          _ttsService!.speak(
+            lastMsg.displayText,
+            voiceKey: voiceKey,
+            messageId: msgId,
+          );
+        }
+
+        // Auto-play: if director mode is active, queue the next character
+        if (_autoPlayActive && _observerMode && _activeGroup != null) {
+          // If TTS is active, wait for it to finish before starting the delay
+          if (_ttsService != null && _ttsService!.isSpeaking) {
+            _waitForTtsThenContinue();
+          } else {
+            final delayMs = (directorDelaySec * 1000).round();
+            Future.delayed(Duration(milliseconds: delayMs), () {
+              if (_autoPlayActive && !_isGenerating) {
+                _autoPlayNext();
+              }
+            });
+          }
+        }
+      }
+
+      // Restore original model if swapped for call mode
+      if (_originalModelName != null && _llmProvider != null) {
+        _llmProvider!.openRouterService.configure(
+          modelName: _originalModelName,
+        );
+      }
+    } catch (e) {
+      final wasCancelled = _cancelRequested;
+      _drainTimer?.cancel();
+      _drainTimer = null;
+      _tokenBuffer.clear();
+      _isGenerating = false;
+      _cancelRequested = false;
+      _generationProgress = 0.0;
+      _isBuffering = false;
+      _generationPhase = GenerationPhase.idle;
+      _prefillStartTime = null;
+      _prefillPromptTokens = 0;
+      _generationStartTime = null;
+
+      // "Connection closed before full header was received" is thrown by the http package
+      // when the HTTP client is closed mid-stream (either by abortGeneration() or a process
+      // crash/restart). Treat it the same as a user cancel — keep the partial response.
+      final errStr = e.toString();
+      final isConnectionClosed =
+          errStr.contains('Connection closed before full header') ||
+          errStr.contains('Connection refused') ||
+          errStr.contains('errno = 61') || // macOS ECONNREFUSED
+          errStr.contains('SocketException') ||
+          (errStr.contains('ClientException') && errStr.contains('closed'));
+      final treatAsCancel = wasCancelled || isConnectionClosed;
+
+      // User-initiated cancel (or forced client close) — keep the partial response, no error message
+      if (treatAsCancel) {
+        // Signal clean completion to SSE listeners
+        _tokenBroadcast.add('__DONE__');
+        if (_sentenceBuffer.trim().isNotEmpty) {
+          _sentenceBroadcast.add(_sentenceBuffer.trim());
+          _sentenceBuffer = '';
+        }
+        _sentenceBroadcast.add('__DONE__');
+
+        // Restore original model if swapped for call mode
+        if (_originalModelName != null && _llmProvider != null) {
+          _llmProvider!.openRouterService.configure(
+            modelName: _originalModelName,
+          );
+        }
+
+        // Save the partial response so regen/continue work
+        await _saveChat();
+        notifyListeners();
+        return;
+      }
+
+      // Build user-friendly error message
+      String errorMsg = e.toString();
+      // Strip Dart's "Exception: " prefix for cleaner display
+      errorMsg = errorMsg.replaceFirst(RegExp(r'^Exception:\s*'), '');
+
+      if (errorMsg.contains('STREAMING_NOT_SUPPORTED') ||
+          errorMsg.contains('HTTP 405')) {
+        errorMsg =
+            'HTTP 405: The server does not support this request. '
+            'If streaming is enabled, try disabling it in Settings > Generation Settings. '
+            'Also verify your API URL is correct.';
+      } else if (errorMsg.contains('Backend process crashed')) {
+        errorMsg =
+            'The backend crashed (likely out of VRAM). '
+            'Try reducing GPU layers or context size in Settings.';
+      } else if (errorMsg.contains('timed out') ||
+          errorMsg.contains('TimeoutException')) {
+        errorMsg =
+            'Request timed out. The model may be too large or the server too slow.';
+      } else if (errorMsg.contains('Connection closed before full header') ||
+          (errorMsg.contains('ClientException') &&
+              errorMsg.contains('closed'))) {
+        errorMsg =
+            'The connection to the backend was closed unexpectedly. '
+            'The model may still be loading — wait for the green ready indicator and try again. '
+            'If this persists, the backend may have run out of VRAM.';
+      }
+
+      _messages.add(
+        ChatMessage(text: errorMsg, sender: "System", isUser: false),
+      );
+
+      // Signal error to SSE listeners
+      _tokenBroadcast.add('__ERROR__');
+
+      // Restore original model if swapped for call mode
+      if (_originalModelName != null && _llmProvider != null) {
+        _llmProvider!.openRouterService.configure(
+          modelName: _originalModelName,
+        );
+      }
+
+      notifyListeners();
+    }
+  }
+}
